feat: 통계를 위한 전체 차량 수 조회

+++ domain/src/main/java/com/pickcar/analytics/presentation/dto/response/StaticAnalyticsResponse.java
@@ -1,10 +1,10 @@
 package com.pickcar.analytics.presentation.dto.response;
 
 public record StaticAnalyticsResponse(
-        Integer totalVehicleCount,
-        Integer reservedVehicleCount,
-        Integer notOperableVehicleCount,
-        Integer expectedReturnCount,
-        Integer delayedCount
+        Long totalVehicleCount,
+        Long reservedVehicleCount,
+        Long notOperableVehicleCount,
+        Long expectedReturnCount,
+        Long delayedCount
 ) {
 }
